shared/kafka_producer.py
# shared/kafka_producer.py
from kafka import KafkaProducer
import json

producer = KafkaProducer(
    bootstrap_servers='localhost:9092',
    value_serializer=lambda v: json.dumps(v).encode('utf-8')
)

def publish(topic, event):
    producer.send(topic, event.to_dict())
    producer.flush()