requirements.txt
fastapi
uvicorn
mysql-connector-python
kafka-python
pydantic
